clean up conversions
