+++ src/compas_rhino/conversions/exceptions.py
