--- src/compas_rhino/geometry/transformations/__init__.py
@@ -1,10 +1,0 @@
-from __future__ import absolute_import
-
-import warnings
-
-warnings.warn(
-    "Conversion functions for transformations have been moved to `compas_rhino.conversions`.",
-    DeprecationWarning,
-)
-
-from compas_rhino.conversions._transformations import *  # noqa : F401 F403
